🐛 fix response - split Acknowledgement XSD
