+++ Dockerfile
@@ -21,7 +21,10 @@ RUN addgroup --gid 2000 --system appgroup && \
     adduser --uid 2000 --system appuser --gid 2000
 
 WORKDIR /app
+RUN chown -R appuser:appgroup /app
+
 COPY --from=builder --chown=appuser:appgroup /app/build/resources/main/xsd/cp/external/ExternalDocumentRequest.xsd /app
+COPY --from=builder --chown=appuser:appgroup /app/build/resources/main/xsd/generic/Acknowledgement/Acknowledgement.xsd /app
 COPY --from=builder --chown=appuser:appgroup /app/build/libs/crime-portal-gateway*.jar /app/app.jar
 COPY --from=builder --chown=appuser:appgroup /app/build/libs/applicationinsights-agent*.jar /app/agent.jar
 COPY --from=builder --chown=appuser:appgroup /app/applicationinsights.json /app
