+++ src/main/resources/application.yml
@@ -65,7 +65,7 @@ ws-sec:
   request-encrypt-actions: "Signature Encrypt"
   response-encrypt-actions: "Signature Encrypt"
   response-signature-parts: "{Element}{http://www.w3.org/2003/05/soap-envelope}Body"
-  response-encryption-parts: "{Content}{http://www.justice.gov.uk/magistrates/external/ExternalDocumentRequest}Acknowledgement"
+  response-encryption-parts: "{Content}{http://www.justice.gov.uk/magistrates/ack}Acknowledgement"
   encryption-sym-algorithm: "http://www.w3.org/2001/04/xmlenc#tripledes-cbc"
 
 included-court-codes: "B01BH,B01CE,B01CN,B01EF,B01GU,B01ND,B03DE,B10BD,B10JJ,B10JQ,B13HT,B14AV,B14ET,B14LO,B16BG,B16HE,B17JA,B20EB,B40IM,B44BA,B47CL,B54MW,B62DC,B62MV"
